Move report- and electrode-related readers
Move report-related readers

Move electrode-related readers

Add new modules
